decode playlists

+++ http-detail.js
@@ -3,9 +3,11 @@
 const net = require('net');
 const tls = require('tls');
 const url = require('url');
+const log = require('./log.js');
 
 function decode(buffer) {
-    var found = buffer.indexOf(Buffer.from([13, 10, 13, 10]));
+    var marker = Buffer.from([13, 10, 13, 10]);
+    var found = buffer.indexOf(marker);
     if (found >= 0) {
         var result = {};
         var singleStr = buffer.toString('ascii', 0, found);
@@ -22,13 +24,17 @@ function decode(buffer) {
         result.statusCode = parseInt(firstLine[1]);
         result.status = firstLine[2];
         result.headers = {};
+        result.content = Buffer.allocUnsafe(buffer.length - found - marker.length);
+        log.debug('found marker at:'+found);
+        log.debug('new buffer length:'+result.content.length);
+        buffer.copy(result.content, 0, found + marker.length, buffer.length);
         for (var i = 1; i < lines.length; i++) {
             var line = lines[i];
             var index = line.indexOf(':');
             if (index < 0) {
-                result.headers[line] = '';
+                result.headers[line.toLowerCase()] = '';
             } else {
-                result.headers[line.substr(0, index)] = line.substr(index + 1);
+                result.headers[line.substr(0, index).toLowerCase()] = line.substr(index + 1).trim();
             }
         }
         return result;
@@ -37,51 +43,68 @@ function decode(buffer) {
 }
 
 function getHeader(u, _options) {
-    console.log('getHeader:'+u);
+    log.debug('getHeader:' + u);
     var options = _options || {};
-    var debug = options.debug || false;
+    var contentsize = options.contentsize || 1000;
+    var checkedWithUser = false;
+    var userWantsData = false;
+    var decoded = null;
+
     return new Promise(function(resolve, reject) {
         var parsed = url.parse(u);
         var buffer = Buffer.alloc(0);
         var resolved = false;
+
         if (parsed.protocol === 'http:' || parsed.protocol === 'https:') {
             var port = null;
             var connect = null;
-            if (parsed.protocol === 'http:'){
+            if (parsed.protocol === 'http:') {
                 connect = net.connect;
                 port = parsed.port || 80;
-            }else if (parsed.protocol === 'https:'){
+            } else if (parsed.protocol === 'https:') {
                 connect = tls.connect;
                 port = parsed.port || 443;
             }
             var client = connect(port, parsed.hostname, function() {
-                var requestStr = 'GET ' + parsed.path + ' HTTP/1.1\n' +
-                    'Host: ' + parsed.hostname + '\n\n';
-                client.write(requestStr);
-                if (debug) {
-                    console.log('connected:\n' + requestStr);
-                }
+                log.debug('Connected to '+parsed.hostname+':'+port);
+                client.setNoDelay(true);
+                var requestStr = 'GET ' + parsed.path + ' HTTP/1.1\r\n' +
+                    'Host: ' + parsed.hostname + '\r\n' +
+                    'User-Agent: RadioBrowser/1.0\r\n' +
+                    'Connection: close\r\n' +
+                    'Accept: */*\r\n\r\n';
+                client.write(requestStr,'ascii');
+                // client.write(Buffer.from([13,10,13,10]));
+                log.debug('Sent to server:\n' + requestStr);
             });
             client.on('data', (data) => {
-                if (debug) {
-                    console.log('connection data length:' + data.length);
-                }
+                log.debug('connection data length:' + data.length);
                 buffer = Buffer.concat([buffer, data]);
-                var decoded = decode(buffer);
-                if (decoded) {
-                    client.destroy();
-                    if (!resolved) {
-                        resolve(decoded);
-                        resolved = true;
+                if (!decoded){
+                    decoded = decode(buffer);
+                    if (decoded) {
+                        if (!checkedWithUser && options.headercheck) {
+                            checkedWithUser = true;
+                            userWantsData = options.headercheck(decoded);
+                        }
+                        if (!userWantsData || decoded.content.length > contentsize) {
+                            client.destroy();
+                            if (!resolved) {
+                                resolve(decoded);
+                                resolved = true;
+                            }
+                        }
                     }
+                } else {
+                    decoded.content = Buffer.concat([decoded.content, data]);
                 }
             });
             client.on('end', () => {
-                if (debug) {
-                    console.log('connection ended');
-                }
+                log.debug('connection ended');
                 if (!resolved) {
-                    var decoded = decode(buffer);
+                    if (!decoded){
+                        decoded = decode(buffer);
+                    }
                     if (decoded) {
                         resolve(decoded);
                     } else {
@@ -90,13 +113,11 @@ function getHeader(u, _options) {
                     resolved = true;
                 }
             });
-            client.on('secureConnect',()=>{
-                console.log('Secure connection established.');
+            client.on('secureConnect', () => {
+                log.debug('Secure connection established.');
             });
             client.on('error', (err) => {
-                if (debug) {
-                    console.log('err:' + err);
-                }
+                log.error('err:' + err);
                 reject(err);
             });
         } else {
@@ -107,8 +128,8 @@ function getHeader(u, _options) {
 
 function getHeaderRecursive(url) {
     return getHeader(url).then((header) => {
-        if (header.headers.Location) {
-            return getHeaderRecursive(header.headers.Location);
+        if (header.headers.location) {
+            return getHeaderRecursive(header.headers.location);
         } else {
             return header;
         }
